requirements.txt
python-telegram-bot==22.5
supabase
httpx
requests
joblib
asyncio
pytz
